Dagger Hilt integrated and Long is used as id for Tasks

+++ android-app/app/src/main/java/com/qadri/to_do/data/OfflineTaskRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 class OfflineTaskRepository(private val taskDao: TaskDao): TaskRepository {
     override fun getAllTask(): Flow<List<Task>> = taskDao.getAllTask()
 
-    override fun getTask(id: Int): Flow<Task> = taskDao.getTask(id)
+    override fun getTask(id: Long): Flow<Task> = taskDao.getTask(id)
 
     override suspend fun insertTask(task: Task) = taskDao.insertTask(task)
